Ya quedo el CRUD

+++ demo/src/main/java/com/example/demo/entity/Especialidad.java
@@ -13,7 +13,7 @@ import jakarta.persistence.*;
 public class Especialidad{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String nombreEspecialidad;
@@ -26,7 +26,7 @@ public class Especialidad{
         this.nombreEspecialidad = nombreEspecialidad;
     }
 
-    public Especialidad(Integer id, String nombreEspecialidad) {
+    public Especialidad(Long id, String nombreEspecialidad) {
         this.id = id;
         this.nombreEspecialidad = nombreEspecialidad;
     }
@@ -35,11 +35,11 @@ public class Especialidad{
 
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
